# refactor: use tree-sitter instead syntect