--- a/Cargo.toml
+++ b/Cargo.toml
@@ -40,7 +40,16 @@ termcolor = "1.1.2"
 html-escape = { version = "0.2.13", optional = true }
 comrak = { version = "0.20.0", optional = true, default-features = false }
 handlebars = { version = "5.0", optional = true, features = ["string_helpers"] }
-syntect = { version = "5.1.0", optional = true, default-features = false, features = [ "parsing", "default-syntaxes", "default-themes", "html", "dump-load", "regex-onig" ] }
+tree-sitter-highlight = { version = "0.20.1", optional = true }
+
+tree-sitter-javascript = { version = "0.20.1", optional = true }
+tree-sitter-typescript = { version = "0.20.3", optional = true }
+tree-sitter-json = { version = "0.20.1", optional = true }
+tree-sitter-regex = { version = "0.20.0", optional = true }
+tree-sitter-css = { version = "0.20.0", optional = true }
+tree-sitter-toml = { version = "0.20.0", optional = true }
+tree-sitter-md = { version = "0.1.7", optional = true }
+
 
 [dev-dependencies]
 anyhow = { version = "1.0.58" }
@@ -50,9 +59,17 @@ tokio = { version = "1.25.0", features = ["full"] }
 pretty_assertions = "1.0.0"
 
 [features]
-default = ["html", "rust"]
+default = ["html", "tree-sitter-all", "rust"]
 rust = []
-html = ["html-escape", "comrak", "handlebars", "syntect"]
+html = ["html-escape", "comrak", "handlebars", "tree-sitter-highlight"]
+tree-sitter-all = ["tree-sitter-javascript", "tree-sitter-typescript", "tree-sitter-json", "tree-sitter-regex", "tree-sitter-css", "tree-sitter-toml", "tree-sitter-md"]
+tree-sitter-javascript = ["dep:tree-sitter-javascript"]
+tree-sitter-typescript = ["dep:tree-sitter-typescript"]
+tree-sitter-json = ["dep:tree-sitter-json"]
+tree-sitter-regex = ["dep:tree-sitter-regex"]
+tree-sitter-css = ["dep:tree-sitter-css"]
+tree-sitter-toml = ["dep:tree-sitter-toml"]
+tree-sitter-md = ["dep:tree-sitter-md"]
 
 [profile.release]
 codegen-units = 1