--- a/src/html/mod.rs
+++ b/src/html/mod.rs
@@ -17,10 +17,13 @@ mod render_context;
 mod search;
 pub mod sidepanels;
 mod symbols;
+mod treesitter;
 mod types;
 mod usage;
 mod util;
 
+pub use self::treesitter::tree_sitter_language_cb;
+pub use self::treesitter::TreeSitterHighlighter;
 pub use jsdoc::ModuleDocCtx;
 pub use pages::generate_symbol_pages_for_module;
 pub use render_context::RenderContext;
@@ -28,16 +31,18 @@ pub use search::generate_search_index;
 pub use symbols::namespace;
 pub use symbols::SymbolContentCtx;
 pub use symbols::SymbolGroupCtx;
+pub use treesitter::CAPTURE_NAMES as TREE_SITTER_HIGHLIGHTER_CAPTURE_NAMES;
 pub use util::compute_namespaced_symbols;
 pub use util::DocNodeKindCtx;
 pub use util::HrefResolver;
 pub use util::NamespacedGlobalSymbols;
 pub use util::UrlResolveKind;
 
-pub const STYLESHEET: &str = include_str!("./templates/styles.css");
+pub const STYLESHEET: &str = include_str!("./templates/styles.gen.css");
 pub const STYLESHEET_FILENAME: &str = "styles.css";
 
-pub const PAGE_STYLESHEET: &str = include_str!("./templates/pages/page.css");
+pub const PAGE_STYLESHEET: &str =
+  include_str!("./templates/pages/page.gen.css");
 pub const PAGE_STYLESHEET_FILENAME: &str = "page.css";
 
 const SEARCH_INDEX_FILENAME: &str = "search_index.js";
@@ -69,7 +74,7 @@ pub struct GenerateCtx<'ctx> {
   pub main_entrypoint: Option<ModuleSpecifier>,
   pub specifiers: Vec<ModuleSpecifier>,
   pub hbs: Handlebars<'ctx>,
-  pub syntect_adapter: comrak_adapters::SyntectAdapter,
+  pub tree_sitter_highlighter: TreeSitterHighlighter,
   pub global_symbols: NamespacedGlobalSymbols,
   pub href_resolver: Rc<dyn HrefResolver>,
   pub rewrite_map: Option<IndexMap<ModuleSpecifier, String>>,
@@ -272,17 +277,8 @@ pub fn setup_hbs<'t>() -> Result<Handlebars<'t>, anyhow::Error> {
   Ok(reg)
 }
 
-pub fn setup_syntect() -> comrak_adapters::SyntectAdapter {
-  let syntax_set: syntect::parsing::SyntaxSet =
-    syntect::dumps::from_uncompressed_data(include_bytes!(
-      "./default_newlines.packdump"
-    ))
-    .unwrap();
-
-  comrak_adapters::SyntectAdapter {
-    syntax_set,
-    theme_set: syntect::highlighting::ThemeSet::load_defaults(),
-  }
+pub fn setup_tree_sitter() -> TreeSitterHighlighter {
+  TreeSitterHighlighter::new(tree_sitter_language_cb)
 }
 
 pub fn generate(
@@ -301,7 +297,7 @@ pub fn generate(
     main_entrypoint: options.main_entrypoint,
     specifiers: doc_nodes_by_url.keys().cloned().collect(),
     hbs: setup_hbs()?,
-    syntect_adapter: setup_syntect(),
+    tree_sitter_highlighter: setup_tree_sitter(),
     global_symbols: options.global_symbols,
     href_resolver: options.href_resolver,
     rewrite_map: options.rewrite_map,