--- a/src/html/templates/anchor.hbs
+++ b/src/html/templates/anchor.hbs
@@ -1,6 +1,6 @@
 <a
   href="#{{id}}"
-  class="anchor float-left leading-none hidden text-stone-600 ml-[-18px] pr-1"
+  class="anchor float-left leading-none hidden text-stone-600 ml-[-24px] p-1 pr-1 pt-1 top-0 bottom-0"
   aria-label="Anchor"
   tabIndex=-1
 >