--- a/src/html/templates/icons/link.hbs
+++ b/src/html/templates/icons/link.hbs
@@ -1,6 +1,6 @@
 <svg
-  width="14"
-  height="14"
+  width="16"
+  height="16"
   viewBox="0 0 14 14"
   fill="none"
   xmlns="http://www.w3.org/2000/svg"