--- a/src/html/templates/source_button.hbs
+++ b/src/html/templates/source_button.hbs
@@ -1,5 +1,5 @@
 <a
-  class="relative rounded p-1.5 cursor-pointer mr-2 hover:bg-stone-300"
+  class="relative rounded p-1.5 cursor-pointer mr-2 border hover:bg-stone-300"
   href="{{this}}">
   {{~> icons/source ~}}
 </a>