--- a/src/html/templates/usage.hbs
+++ b/src/html/templates/usage.hbs
@@ -1,13 +1,10 @@
-<div class="usage">
-  <div class="relative">
-    {{{~import_statement~}}}
-
-    <div class="absolute top-0 right-2 flex items-center h-full">
-      <button
-        class="usage_button rounded p-1.5 cursor-pointer hover:bg-stone-300 hover:before:block hover:after:block after:content-['Click_to_copy'] focus:after:content-['Copied!'] focus:after:w-[70px] focus:after:left-[-20px]"
-        onclick='navigator?.clipboard?.writeText({{raw_import_statement}})'>
-        {{~> icons/copy ~}}
-      </button>
-    </div>
+<div class="usage relative">
+  {{~import_statement~}}
+  <div class="absolute top-0 right-2 flex items-center h-full">
+    <button
+      class="usage_button rounded p-1.5 cursor-pointer border bg-stone-50 hover:bg-stone-300 hover:before:block hover:after:block after:content-['Click_to_copy'] focus:after:content-['Copied!'] focus:after:w-[70px] focus:after:left-[-20px]"
+      onclick='navigator?.clipboard?.writeText({{raw_import_statement}})'>
+      {{~> icons/copy ~}}
+    </button>
   </div>
 </div>