--- a/src/html/treesitter.rs
+++ b/src/html/treesitter.rs
@@ -115,6 +115,7 @@ pub fn tree_sitter_language_cb(
       "md" | "markdown" => tree_sitter_language_markdown(),
       "toml" => tree_sitter_language_toml(),
       "regex" => tree_sitter_language_regex(),
+      "rs" | "rust" => tree_sitter_language_rust(),
       _ => continue,
     };
     return Some(cfg);
@@ -278,6 +279,21 @@ fn tree_sitter_language_regex() -> &'static HighlightConfiguration {
       "",
     )
     .expect("failed to initialize tree_sitter_regex highlighter");
+    config.configure(CAPTURE_NAMES);
+    config
+  })
+}
+
+fn tree_sitter_language_rust() -> &'static HighlightConfiguration {
+  static CONFIG: OnceLock<HighlightConfiguration> = OnceLock::new();
+  CONFIG.get_or_init(|| {
+    let mut config = HighlightConfiguration::new(
+      tree_sitter_rust::language(),
+      tree_sitter_rust::HIGHLIGHT_QUERY,
+      tree_sitter_rust::INJECTIONS_QUERY,
+      "",
+    )
+    .expect("failed to initialize tree_sitter_rust highlighter");
     config.configure(CAPTURE_NAMES);
     config
   })