--- a/tailwind.config.ts
+++ b/tailwind.config.ts
@@ -42,6 +42,5 @@ export default {
       colors: extraColors,
     },
   },
-  important: ".ddoc",
   darkMode: "class",
 } as Config;