--- a/tests/html_test.rs
+++ b/tests/html_test.rs
@@ -305,7 +305,7 @@ async fn symbol_group() {
     .join("symbol_group.json");
 
   // uncomment to regenerate symbol_group.json
-  //std::fs::write(&symbol_group_json_path, &files_json);
+  // std::fs::write(&symbol_group_json_path, &files_json);
 
   let symbol_group_json = read_to_string(symbol_group_json_path).unwrap();