tree-sitter is a much nicer parser, and pulls in much fewer deps.

Also I've done some minor cleanups on the generated HTML, and a refactor how the CSS is generated to further shrink it (passing it through lightningcss).