@lucacasonato 2024-01-15
@crowlKats It has much more accurate output than any regexp based syntax highlighter, because it uses a "proper" parser rather than a regexp matcher. It can generally offer the same accuracy as semantic syntax highlighting (which usually requires a pass through TSC or similar LSP server) but at or exceeding the performance of regexp based syntax highlighter.

tree-sitter is also the syntax highlighter used by GitHub to highlight code for most languages.

Having this in the binary would also allow us to implement go-to-definition for JS / TS files in the registry source view at some point.

|  | Before | After |
| - | - | - |
| ddoc size | 11M | 13M |
| ddoc --html types.d.ts | 1.985 s ±  0.036 s | 1.872 s ±  0.046 s |
| generated_docs size | 69M | 71MB |

The generated_docs size difference mainly boils down to the examples now having more accurate highlighting, for example the idents in a member expression are now highlighted correctly:

Before
![image](https://github.com/denoland/deno_doc/assets/7829205/434c3896-7679-4c44-a17e-e7423ec99543)

After:
![image](https://github.com/denoland/deno_doc/assets/7829205/59839a71-b7b8-41c1-b165-1254c06625a5)